libraries/logging/pyproject.toml
[tool.poetry]
name = "logging"
version = "0.1.0"
description = ""
authors = ["Sander Möller <sander.mollerr@gmail.com>"]
packages = [
    { include = "ns", from = "src" }
]

[tool.poetry.dependencies]
python = "3.12"
python-json-logger = "^2.0.7"
opentelemetry-distro = {version="^0.41b0", optional=true}

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"

[tool.poetry.extras]
otel = ["opentelemetry-distro"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
